moved DeviceTypeMapper to shared/config
