+++ app/src/main/kotlin/com/wire/android/feature/auth/client/Client.kt
@@ -1,8 +1,8 @@
 package com.wire.android.feature.auth.client
 
 import com.wire.android.core.config.DeviceClass
+import com.wire.android.core.config.DeviceType
 import com.wire.android.core.crypto.model.PreKey
-import com.wire.android.feature.auth.client.datasource.DeviceType
 
 data class Client(
     val id: String,
